+++ spec/support/active_record.rb
@@ -8,6 +8,7 @@ end
 
 ActiveRecord::Base.connection.execute("DROP TABLE IF EXISTS 'countries'")
 ActiveRecord::Base.connection.execute("DROP TABLE IF EXISTS 'cities'")
+ActiveRecord::Base.connection.execute("DROP TABLE IF EXISTS 'locations'")
 ActiveRecord::Schema.define do
   create_table :countries do |t|
     t.column :name, :string
@@ -19,8 +20,16 @@ ActiveRecord::Schema.define do
   create_table :cities do |t|
     t.column :country_id, :integer
     t.column :name, :string
+    t.column :historical_name, :string
+    t.column :description, :string
     t.column :rating, :integer
     t.column :updated_at, :datetime
+  end
+
+  create_table :locations do |t|
+    t.column :city_id, :integer
+    t.column :lat, :string
+    t.column :lon, :string
   end
 end